FIR/UAST: commonize local function conversion
