+++ uast/uast-kotlin-base/src/org/jetbrains/uast/kotlin/expressions/KotlinUReturnExpression.kt
@@ -7,7 +7,6 @@ package org.jetbrains.uast.kotlin
 
 import org.jetbrains.kotlin.psi.KtReturnExpression
 import org.jetbrains.uast.*
-import org.jetbrains.uast.kotlin.expressions.KotlinLocalFunctionULambdaExpression
 
 class KotlinUReturnExpression(
     override val sourcePsi: KtReturnExpression,
