+++ uast/uast-kotlin-fir/testData/legacyIdentifiers/Reified.identifiers.fir.txt
@@ -2,22 +2,22 @@ test -> USimpleNameReferenceExpression (identifier = test)
 pkg -> USimpleNameReferenceExpression (identifier = pkg)
 Context -> UClass (name = Context)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-ownSystemService1 -> [!] UnknownKotlinExpression (FUN)
+ownSystemService1 -> UDeclarationsExpression
 getSystemService -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
 T -> USimpleNameReferenceExpression (identifier = T)
 java -> USimpleNameReferenceExpression (identifier = java)
-ownSystemService2 -> [!] UnknownKotlinExpression (FUN)
+ownSystemService2 -> UDeclarationsExpression
 getSystemService -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
 String -> USimpleNameReferenceExpression (identifier = String)
 java -> USimpleNameReferenceExpression (identifier = java)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 Context -> USimpleNameReferenceExpression (identifier = Context)
-systemService1 -> [!] UnknownKotlinExpression (FUN)
+systemService1 -> UDeclarationsExpression
 getSystemService -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
 T -> USimpleNameReferenceExpression (identifier = T)
 java -> USimpleNameReferenceExpression (identifier = java)
 Context -> USimpleNameReferenceExpression (identifier = Context)
-systemService2 -> [!] UnknownKotlinExpression (FUN)
+systemService2 -> UDeclarationsExpression
 getSystemService -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
 String -> USimpleNameReferenceExpression (identifier = String)
 java -> USimpleNameReferenceExpression (identifier = java)
