+++ uast/uast-kotlin-fir/testData/legacyIdentifiers/ReifiedParameters.identifiers.fir.txt
@@ -1,5 +1,5 @@
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-functionWithLambda -> [!] UnknownKotlinExpression (FUN)
+functionWithLambda -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 process -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -9,7 +9,7 @@ Int -> USimpleNameReferenceExpression (identifier = Int)
 process -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-functionWithVararg -> [!] UnknownKotlinExpression (FUN)
+functionWithVararg -> UDeclarationsExpression
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 Int -> USimpleNameReferenceExpression (identifier = Int)
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -18,27 +18,27 @@ T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
-functionWithReceiver -> [!] UnknownKotlinExpression (FUN)
+functionWithReceiver -> UDeclarationsExpression
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 Int -> USimpleNameReferenceExpression (identifier = Int)
 T -> USimpleNameReferenceExpression (identifier = T)
 this -> USimpleNameReferenceExpression (identifier = this)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
-`name with spaces` -> [!] UnknownKotlinExpression (FUN)
+`name with spaces` -> UDeclarationsExpression
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 Int -> USimpleNameReferenceExpression (identifier = Int)
 T -> USimpleNameReferenceExpression (identifier = T)
 this -> USimpleNameReferenceExpression (identifier = this)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-functionWithParamAnnotation -> [!] UnknownKotlinExpression (FUN)
+functionWithParamAnnotation -> UDeclarationsExpression
 Suppress -> USimpleNameReferenceExpression (identifier = Suppress)
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-functionUnresolved -> [!] UnknownKotlinExpression (FUN)
+functionUnresolved -> UDeclarationsExpression
 Suppress -> USimpleNameReferenceExpression (identifier = Suppress)
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 Unresolved -> USimpleNameReferenceExpression (identifier = Unresolved)
