+++ uast/uast-kotlin-fir/testData/legacyIdentifiers/ReifiedReturnType.identifiers.fir.txt
@@ -1,5 +1,5 @@
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-function1 -> [!] UnknownKotlinExpression (FUN)
+function1 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -7,7 +7,7 @@ Int -> USimpleNameReferenceExpression (identifier = Int)
 s -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 String -> USimpleNameReferenceExpression (identifier = String)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-function2 -> [!] UnknownKotlinExpression (FUN)
+function2 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -18,7 +18,7 @@ T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 CharSequence -> USimpleNameReferenceExpression (identifier = CharSequence)
-function2CharSequence -> [!] UnknownKotlinExpression (FUN)
+function2CharSequence -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -28,7 +28,7 @@ String -> USimpleNameReferenceExpression (identifier = String)
 T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-function3 -> [!] UnknownKotlinExpression (FUN)
+function3 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -36,7 +36,7 @@ Int -> USimpleNameReferenceExpression (identifier = Int)
 s -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 String -> USimpleNameReferenceExpression (identifier = String)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-function4 -> [!] UnknownKotlinExpression (FUN)
+function4 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -46,7 +46,7 @@ String -> USimpleNameReferenceExpression (identifier = String)
 T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-function5 -> [!] UnknownKotlinExpression (FUN)
+function5 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -57,7 +57,7 @@ Int -> USimpleNameReferenceExpression (identifier = Int)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 Activity -> USimpleNameReferenceExpression (identifier = Activity)
 T -> USimpleNameReferenceExpression (identifier = T)
-function6 -> [!] UnknownKotlinExpression (FUN)
+function6 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -67,7 +67,7 @@ String -> USimpleNameReferenceExpression (identifier = String)
 T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-function7 -> [!] UnknownKotlinExpression (FUN)
+function7 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -77,7 +77,7 @@ String -> USimpleNameReferenceExpression (identifier = String)
 T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-function8 -> [!] UnknownKotlinExpression (FUN)
+function8 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -87,7 +87,7 @@ String -> USimpleNameReferenceExpression (identifier = String)
 T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-function9 -> [!] UnknownKotlinExpression (FUN)
+function9 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -97,7 +97,7 @@ String -> USimpleNameReferenceExpression (identifier = String)
 T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
-function10 -> [!] UnknownKotlinExpression (FUN)
+function10 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -108,7 +108,7 @@ T -> USimpleNameReferenceExpression (identifier = T)
 t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
-function11 -> [!] UnknownKotlinExpression (FUN)
+function11 -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -120,7 +120,7 @@ t -> USimpleNameReferenceExpression (identifier = t)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 CharSequence -> USimpleNameReferenceExpression (identifier = CharSequence)
 T -> USimpleNameReferenceExpression (identifier = T)
-function11CharSequence -> [!] UnknownKotlinExpression (FUN)
+function11CharSequence -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -134,7 +134,7 @@ CharSequence -> USimpleNameReferenceExpression (identifier = CharSequence)
 B -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
 T -> USimpleNameReferenceExpression (identifier = T)
-function12CharSequence -> [!] UnknownKotlinExpression (FUN)
+function12CharSequence -> UDeclarationsExpression
 t -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
 B -> USimpleNameReferenceExpression (identifier = B)
 i -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
@@ -166,6 +166,6 @@ Foo -> UClass (name = Foo)
 T -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 Z -> [!] UnknownKotlinExpression (TYPE_PARAMETER)
 T -> USimpleNameReferenceExpression (identifier = T)
-foo -> [!] UnknownKotlinExpression (FUN)
+foo -> UDeclarationsExpression
 Z -> USimpleNameReferenceExpression (identifier = Z)
     TODO -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
