+++ uast/uast-kotlin-fir/testData/legacyRenderLog/LambdaReturn.log.fir.txt
@@ -14,10 +14,34 @@ UFile (package = org.jetbrains.uast.kotlin)
                             [!] UnknownKotlinExpression (LAMBDA_EXPRESSION)
                 UDeclarationsExpression
                     ULocalVariable (name = lam4)
-                        [!] UnknownKotlinExpression (FUN)
+                        ULambdaExpression
+                            UParameter (name = a)
+                                UAnnotation (fqName = org.jetbrains.annotations.NotNull)
+                            UBlockExpression
+                                UIfExpression
+                                    UBinaryExpression (operator = <)
+                                        USimpleNameReferenceExpression (identifier = a)
+                                        ULiteralExpression (value = 5)
+                                    UReturnExpression
+                                        ULiteralExpression (value = "5")
+                                UIfExpression
+                                    UBinaryExpression (operator = >)
+                                        USimpleNameReferenceExpression (identifier = a)
+                                        ULiteralExpression (value = 0)
+                                    UReturnExpression
+                                        ULiteralExpression (value = "1")
+                                    UReturnExpression
+                                        ULiteralExpression (value = "2")
                 UDeclarationsExpression
                     ULocalVariable (name = lam5)
-                        [!] UnknownKotlinExpression (FUN)
+                        ULambdaExpression
+                            UParameter (name = a)
+                                UAnnotation (fqName = org.jetbrains.annotations.NotNull)
+                            UBlockExpression
+                                UReturnExpression
+                                    UBinaryExpression (operator = +)
+                                        ULiteralExpression (value = "a")
+                                        USimpleNameReferenceExpression (identifier = a)
                 [!] UnknownKotlinExpression (CALL_EXPRESSION)
                 UDeclarationsExpression
                     ULocalVariable (name = x)
