+++ uast/uast-kotlin-fir/testData/legacyRenderLog/LambdaReturn.render.fir.txt
@@ -5,8 +5,13 @@ public final class LambdaReturnKt {
         var lam1: <ErrorType> = [!] UnknownKotlinExpression (LAMBDA_EXPRESSION)
         var lam2: <ErrorType> = [!] UnknownKotlinExpression (LAMBDA_EXPRESSION)
         var lam3: <ErrorType> = lbd@ [!] UnknownKotlinExpression (LAMBDA_EXPRESSION)
-        var lam4: <ErrorType> = [!] UnknownKotlinExpression (FUN)
-        var lam5: <ErrorType> = [!] UnknownKotlinExpression (FUN)
+        var lam4: <ErrorType> = fun (@org.jetbrains.annotations.NotNull var a: <ErrorType>) {
+                if (a < 5) return "5"
+                if (a > 0) return "1" else return "2"
+        }
+        var lam5: <ErrorType> = fun (@org.jetbrains.annotations.NotNull var a: <ErrorType>) {
+                return "a" + a
+        }
         [!] UnknownKotlinExpression (CALL_EXPRESSION)
         var x: <ErrorType> = [!] UnknownKotlinExpression (LAMBDA_EXPRESSION)
         var y: <ErrorType> = [!] UnknownKotlinExpression (LAMBDA_EXPRESSION)
