+++ uast/uast-kotlin-fir/testData/legacyRenderLog/Lambdas.log.fir.txt
@@ -15,7 +15,9 @@ UFile (package = )
                 [!] UnknownKotlinExpression (CALL_EXPRESSION)
                 UDeclarationsExpression
                     ULocalVariable (name = baz)
-                        [!] UnknownKotlinExpression (FUN)
+                        ULambdaExpression
+                            UBlockExpression
+                                [!] UnknownKotlinExpression (CALL_EXPRESSION)
                 [!] UnknownKotlinExpression (CALL_EXPRESSION)
         UMethod (name = lambdaInPlaceCall)
             UBlockExpression
