+++ uast/uast-kotlin-fir/testData/legacyRenderLog/Lambdas.render.fir.txt
@@ -6,7 +6,9 @@ public final class LambdasKt {
     }
     public static final fun doSelectItem(@null selectItemFunction: kotlin.jvm.functions.Function0<? extends kotlin.Unit>) : void {
         [!] UnknownKotlinExpression (CALL_EXPRESSION)
-        var baz: <ErrorType> = [!] UnknownKotlinExpression (FUN)
+        var baz: <ErrorType> = fun () {
+                [!] UnknownKotlinExpression (CALL_EXPRESSION)
+        }
         [!] UnknownKotlinExpression (CALL_EXPRESSION)
     }
     public static final fun lambdaInPlaceCall() : void {
