+++ uast/uast-kotlin-fir/testData/legacyRenderLog/LocalDeclarations.render.fir.txt
@@ -3,9 +3,15 @@ public final class LocalDeclarationsKt {
         public static final class Local {
             public fun Local() = UastEmptyExpression
         }
-        [!] UnknownKotlinExpression (FUN)
-        var baz: <ErrorType> = [!] UnknownKotlinExpression (FUN)
-        [!] UnknownKotlinExpression (FUN)
+        var bar: <ErrorType> = fun () {
+                return [!] UnknownKotlinExpression (CALL_EXPRESSION)
+        }
+        var baz: <ErrorType> = fun () {
+                [!] UnknownKotlinExpression (CALL_EXPRESSION)
+        }
+        var someLocalFun: <ErrorType> = fun (@null var text: <ErrorType>) {
+                return 42
+        }
         return [!] UnknownKotlinExpression (CALL_EXPRESSION) == [!] UnknownKotlinExpression (CALL_EXPRESSION)
     }
 }
