+++ uast/uast-kotlin-fir/testData/legacyTypes/Lambdas.types.fir.txt
@@ -13,9 +13,11 @@ UFile (package = ) [import java.util.stream.Stream...]
                 UAnnotation (fqName = null) [@null]
             UBlockExpression [{...}] : PsiType:NonExistentClass
                 [!] UnknownKotlinExpression (CALL_EXPRESSION) [[!] UnknownKotlinExpression (CALL_EXPRESSION)]
-                UDeclarationsExpression [var baz: <ErrorType> = [!] UnknownKotlinExpression (FUN)]
-                    ULocalVariable (name = baz) [var baz: <ErrorType> = [!] UnknownKotlinExpression (FUN)]
-                        [!] UnknownKotlinExpression (FUN) [[!] UnknownKotlinExpression (FUN)]
+                UDeclarationsExpression [var baz: <ErrorType> = fun () {...}]
+                    ULocalVariable (name = baz) [var baz: <ErrorType> = fun () {...}]
+                        ULambdaExpression [fun () {...}]
+                            UBlockExpression [{...}] : PsiType:NonExistentClass
+                                [!] UnknownKotlinExpression (CALL_EXPRESSION) [[!] UnknownKotlinExpression (CALL_EXPRESSION)]
                 [!] UnknownKotlinExpression (CALL_EXPRESSION) [[!] UnknownKotlinExpression (CALL_EXPRESSION)]
         UMethod (name = lambdaInPlaceCall) [public static final fun lambdaInPlaceCall() : void {...}]
             UBlockExpression [{...}] : PsiType:Unit
