+++ uast/uast-kotlin-fir/testData/legacyTypes/LocalDeclarations.types.fir.txt
@@ -5,11 +5,25 @@ UFile (package = ) [public final class LocalDeclarationsKt {...]
                 UDeclarationsExpression [public static final class Local {...}]
                     UClass (name = Local) [public static final class Local {...}]
                         UMethod (name = Local) [public fun Local() = UastEmptyExpression]
-                [!] UnknownKotlinExpression (FUN) [[!] UnknownKotlinExpression (FUN)]
-                UDeclarationsExpression [var baz: <ErrorType> = [!] UnknownKotlinExpression (FUN)]
-                    ULocalVariable (name = baz) [var baz: <ErrorType> = [!] UnknownKotlinExpression (FUN)]
-                        [!] UnknownKotlinExpression (FUN) [[!] UnknownKotlinExpression (FUN)]
-                [!] UnknownKotlinExpression (FUN) [[!] UnknownKotlinExpression (FUN)]
+                UDeclarationsExpression [var bar: <ErrorType> = fun () {...}]
+                    UVariable (name = bar) [var bar: <ErrorType> = fun () {...}]
+                        ULambdaExpression [fun () {...}]
+                            UBlockExpression [{...}]
+                                UReturnExpression [return [!] UnknownKotlinExpression (CALL_EXPRESSION)]
+                                    [!] UnknownKotlinExpression (CALL_EXPRESSION) [[!] UnknownKotlinExpression (CALL_EXPRESSION)]
+                UDeclarationsExpression [var baz: <ErrorType> = fun () {...}]
+                    ULocalVariable (name = baz) [var baz: <ErrorType> = fun () {...}]
+                        ULambdaExpression [fun () {...}]
+                            UBlockExpression [{...}] : PsiType:Local
+                                [!] UnknownKotlinExpression (CALL_EXPRESSION) [[!] UnknownKotlinExpression (CALL_EXPRESSION)]
+                UDeclarationsExpression [var someLocalFun: <ErrorType> = fun (@null var text: <ErrorType>) {...}]
+                    UVariable (name = someLocalFun) [var someLocalFun: <ErrorType> = fun (@null var text: <ErrorType>) {...}]
+                        ULambdaExpression [fun (@null var text: <ErrorType>) {...}]
+                            UParameter (name = text) [@null var text: <ErrorType>]
+                                UAnnotation (fqName = null) [@null]
+                            UBlockExpression [{...}]
+                                UReturnExpression [return 42]
+                                    ULiteralExpression (value = 42) [42] : PsiType:int
                 UReturnExpression [return [!] UnknownKotlinExpression (CALL_EXPRESSION) == [!] UnknownKotlinExpression (CALL_EXPRESSION)] : PsiType:Void
                     UBinaryExpression (operator = ==) [[!] UnknownKotlinExpression (CALL_EXPRESSION) == [!] UnknownKotlinExpression (CALL_EXPRESSION)] : PsiType:boolean
                         [!] UnknownKotlinExpression (CALL_EXPRESSION) [[!] UnknownKotlinExpression (CALL_EXPRESSION)]
